requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
psycopg2-binary==2.9.9
pydantic==2.9.0
